fix: review fixes again
