+++ projects/demo/src/modules/components/progress-bar/examples/2/index.html
@@ -18,9 +18,7 @@
 <p>
     Set component's input property
     <code>color</code>
-    to get more complex color combinations. Use directive
-    <code>tuiProgressFixedGradient</code>
-    to make gradient fixed.
+    to get more complex color combinations.
 </p>
 
 <progress
@@ -30,6 +28,12 @@
     class="progress"
     [value]="fastValue$ | async"
 ></progress>
+
+<p>
+    Use directive
+    <code>tuiProgressFixedGradient</code>
+    to make gradient fixed.
+</p>
 
 <progress
     color="linear-gradient(to right, var(--tui-chart-categorical-02), var(--tui-chart-categorical-14), var(--tui-chart-categorical-12))"
