+++ projects/kit/components/progress/index.ts
@@ -2,7 +2,7 @@ export * from './progress';
 export * from './progress.options';
 export * from './progress-bar/progress-bar.component';
 export * from './progress-bar/progress-color-segments.directive';
-export * from './progress-bar/progress-fixed-gradient.directive';
+export * from './progress-bar/fixed-gradient/progress-fixed-gradient.directive';
 export * from './progress-circle/progress-circle.component';
 export * from './progress-label/progress-label.component';
 export * from './progress-segmented/progress-segmented.directive';
