+++ projects/kit/styles/components/progress-bar.less
@@ -104,18 +104,6 @@
         border-radius: inherit;
     }
 
-    &[tuiProgressFixedGradient] {
-        .progressValue({
-            border-radius: inherit;
-
-            width: 100% !important;
-            margin-right: calc(-100% + var(--tui-progress-percent));
-
-            clip-path: inset(0 calc(100% - var(--tui-progress-percent)) 0 0 round var(--tui-radius-m));
-            transition: all var(--tui-duration) linear;
-        });
-    }
-
     &::-webkit-progress-bar {
         background: transparent;
         border-radius: inherit;
